// src/core/interfaces/data/asset.interface.tsx
export interface InitialValues {
  asset_type: string;
  asset_id: string;
  co_code: string;
  company: string;
  sub_type: string;
  category: string;
  audit_category: string;
  vendor: string;
  purchase_date: string;
  purchase_no: string;
  purchase_quantity: string;
  purchase_value: string;
  description: string;
  product: string;
  branch: string;
  supervisor: string;
  employee: string;
  arn_no: string;
  arn_date: string;
  asset_duplicate_id: string;
  counting_remark: string;
  condition: string;
  condition_date: string;
  last_check_date: string;
  documents: string[];
}